requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.0.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djongo==1.3.3
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4
toml==0.10.1
